add alien model and framework for setting arm poses, still needs a bit of polishing but almost ready to place aliens around the scene

+++ include/UFO.h
@@ -13,7 +13,7 @@ class UFO : public Drawable {
         void updateSmoke();
         void setDishAngle(float angle) { dishAngle = angle; };
         void incDishAngle();
-        void incTakeOffHeight(float keyframe);
+        void incTakeOffHeight(int keyframe);
         void takeOff();
 
     private:
